+++ src/skt_tools.cpp
@@ -927,6 +927,9 @@ bool Sketch_tools::clear_tmps()
   m_bone_centers.reset();
   m_bone_r1.reset();
   m_bone_r2.reset();
+  m_bone_waist.reset();
+  m_bone_hole_r1.reset();
+  m_bone_hole_r2.reset();
   clear_all(m_tmp_node_idxs, m_tmp_shp, m_tmp_edges);
   m_sketch.m_dims.on_clear_tmps();